third_party/blink/web_tests/wpt_internal/web-preferences-api/reducedTransparency.https.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>WebPreferences Test: reducedTransparency attribute tests</title>
    <link rel="author" title="Luke Warlow" href="mailto:luke@warlow.dev" />
    <link rel="help" href="https://wicg.github.io/web-preferences-api/" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(() => {
        assert_true('preferences' in navigator, 'navigator.preferences defined');
        assert_true('reducedTransparency' in navigator.preferences, 'navigator.preferences.reducedTransparency defined');
      }, "navigator.preferences.reducedTransparency supported");
      test(() => {
        assert_equals(navigator.preferences.reducedTransparency.override, null);
      }, "navigator.preferences.reducedTransparency.override is null by default");
      test(() => {
        assert_true('validValues' in navigator.preferences.reducedTransparency, 'navigator.preferences.reducedTransparency.validValues defined');
        assert_true(navigator.preferences.reducedTransparency.validValues.includes('reduce'), 'navigator.preferences.reducedTransparency.validValues includes "reduce"');
        assert_false(navigator.preferences.reducedTransparency.validValues.includes('no-preference'), 'navigator.preferences.reducedTransparency.validValues excludes "no-preference"');
      }, "navigator.preferences.reducedTransparency.validValues supported");
      promise_test((t) => {
        return navigator.preferences.reducedTransparency.requestOverride('this-is-an-invalid-value')
          .then(() => {
            assert_not_reached('requestOverride with invalid value should not resolve');
          })
          .catch((e) => {
            assert_true(e instanceof DOMException);
            assert_equals(navigator.preferences.reducedTransparency.override, null);
            t.done();
          });
      }, "navigator.preferences.reducedTransparency.requestOverride() with invalid value throws");
      promise_test((t) => {
        return navigator.preferences.reducedTransparency.requestOverride('no-preference')
          .then(() => {
            assert_not_reached('requestOverride with invalid value should not resolve');
          })
          .catch((e) => {
            assert_true(e instanceof DOMException);
            assert_equals(navigator.preferences.reducedTransparency.override, null);
            t.done();
          });
      }, "navigator.preferences.reducedTransparency.requestOverride('no-preference') throws");
      promise_test(async (t) => {
        await navigator.preferences.reducedTransparency.requestOverride('reduce');
        assert_equals(navigator.preferences.reducedTransparency.override, 'reduce');
        navigator.preferences.reducedTransparency.clearOverride();
        assert_equals(navigator.preferences.reducedTransparency.override, null);
        t.done();
      }, "navigator.preferences.reducedTransparency.clearOverride() works");
      promise_test(async (t) => {
        await navigator.preferences.reducedTransparency.requestOverride('reduce');
        assert_equals(navigator.preferences.reducedTransparency.override, 'reduce');
        await navigator.preferences.reducedTransparency.requestOverride(null);
        assert_equals(navigator.preferences.reducedTransparency.override, null);
        t.done();
      }, "navigator.preferences.reducedTransparency.requestOverride(null) works like clearOverride()");
      promise_test(async (t) => {
        await navigator.preferences.reducedTransparency.requestOverride('reduce');
        assert_equals(navigator.preferences.reducedTransparency.override, 'reduce');
        await navigator.preferences.reducedTransparency.requestOverride('');
        assert_equals(navigator.preferences.reducedTransparency.override, null);
        t.done();
      }, "navigator.preferences.reducedTransparency.requestOverride('') works like clearOverride()");
      promise_test(async (t) => {
        await navigator.preferences.reducedTransparency.requestOverride('reduce');
        assert_true(window.matchMedia('(prefers-reduced-transparency: reduce)').matches, '(prefers-reduced-transparency: reduce) matches');
        navigator.preferences.reducedTransparency.clearOverride();
        assert_true(window.matchMedia('(prefers-reduced-transparency: no-preference)').matches);
        t.done();
      }, "navigator.preferences.reducedTransparency.requestOverride() updates (prefers-reduced-transparency)");
    </script>
  </body>
</html>
